--- a/vm/channel.go
+++ b/vm/channel.go
@@ -138,9 +138,10 @@ func builtinChannelInstanceMethods() []*BuiltinMethodObject {
 			// @return [Null]
 			Name: "close",
 			Fn: func(receiver Object, sourceLine int, t *Thread, args []Object, blockFrame *normalCallFrame) Object {
-				if e, aLen := 0, len(args); e != aLen {
-					return t.vm.InitErrorObject(errors.ArgumentError, sourceLine, errors.WrongNumberOfArgument, e, aLen)
+				if len(args) != 0 {
+					return t.vm.InitErrorObject(errors.ArgumentError, sourceLine, errors.WrongNumberOfArgument, 0, len(args))
 				}
+				
 				c := receiver.(*ChannelObject)
 
 				if c.ChannelState == chClosed {
@@ -179,9 +180,10 @@ func builtinChannelInstanceMethods() []*BuiltinMethodObject {
 			// @return [Object]
 			Name: "deliver",
 			Fn: func(receiver Object, sourceLine int, t *Thread, args []Object, blockFrame *normalCallFrame) Object {
-				if e, aLen := 1, len(args); e != aLen {
-					return t.vm.InitErrorObject(errors.ArgumentError, sourceLine, errors.WrongNumberOfArgument, e, aLen)
+				if len(args) != 1 {
+					return t.vm.InitErrorObject(errors.ArgumentError, sourceLine, errors.WrongNumberOfArgument, 1, len(args))
 				}
+				
 				c := receiver.(*ChannelObject)
 
 				if c.ChannelState == chClosed {
@@ -217,9 +219,10 @@ func builtinChannelInstanceMethods() []*BuiltinMethodObject {
 			// @return [Object]
 			Name: "receive",
 			Fn: func(receiver Object, sourceLine int, t *Thread, args []Object, blockFrame *normalCallFrame) Object {
-				if e, aLen := 0, len(args); e != aLen {
-					return t.vm.InitErrorObject(errors.ArgumentError, sourceLine, errors.WrongNumberOfArgument, e, aLen)
+				if len(args) != 0 {
+					return t.vm.InitErrorObject(errors.ArgumentError, sourceLine, errors.WrongNumberOfArgument, 0, len(args))
 				}
+				
 				c := receiver.(*ChannelObject)
 
 				if c.ChannelState == chClosed {