Unifying the error messages for length checking in the rest of the classes.

```
benchmarking f54dc3e582496de9dd459c3858a38ff3844c3f54
benchmarking 3a8ae31e9ef7e1bed0007c9d15a7e4374625ab7e
x86_64-darwin17
benchmark                              old ns/op     new ns/op     delta
BenchmarkBasicMath/add-8               3239          3234          -0.15%
BenchmarkBasicMath/subtract-8          3158          3287          +4.08%
BenchmarkBasicMath/multiply-8          3250          3152          -3.02%
BenchmarkBasicMath/divide-8            3337          3175          -4.85%
BenchmarkConcurrency/concurrency-8     7862254       7397854       -5.91%

benchmark                              old allocs     new allocs     delta
BenchmarkBasicMath/add-8               33             33             +0.00%
BenchmarkBasicMath/subtract-8          33             33             +0.00%
BenchmarkBasicMath/multiply-8          33             33             +0.00%
BenchmarkBasicMath/divide-8            33             33             +0.00%
BenchmarkConcurrency/concurrency-8     53465          53516          +0.10%

benchmark                              old bytes     new bytes     delta
BenchmarkBasicMath/add-8               1568          1568          +0.00%
BenchmarkBasicMath/subtract-8          1568          1568          +0.00%
BenchmarkBasicMath/multiply-8          1568          1568          +0.00%
BenchmarkBasicMath/divide-8            1568          1568          +0.00%
BenchmarkConcurrency/concurrency-8     2725831       2729690       +0.14%
Switched to branch 'unify_argument_errors_cont'
```

If it is OK, I'll squash and merge them.