@hachi8833 2018-08-25
Is it OK to merge? > @st0012